Adding BaseRebalanceSegmentStrategy abstract class to encapsulate common methods such as updateIdealState
